Clarify some variable names and methods
The deconv_wiener method and its arguments was very confusing since
not only the filtering but also the next convolution was performed
under the hood. For that reason, I decided to separate that method
into two methods, the deconv_wiener working with the source psf
and the homogenization kernel dealing with actually producing the
kernel.

Also some variable names were clarified like tk replaced with kernel.